Minor changes to code
- Changed the type of Expression's getAnswer() to double
- Switched the input letters for the menu so they are accurate
- Changed the type return of evaluate_expression() to double
- Modified evaluate_expression() to return NaN if syntax check fails
- Slightly modified the output of the menu to place syntax errors at the
right place
- Expanded the number of expressions that can be read from a file from
50 to 1000

+++ CS303_Project2/CS303_Project2/Evaluator.h
@@ -4,6 +4,6 @@
 
 	//Evaluate expression
 	//Takes a string expression and evalutates it solution
-	int evaluate_expression(string &input);
+	double evaluate_expression(string &input);
